+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.0] - 2024-06-06
+- Added OpenTofu as an option to run the commands beyond the default terraform.
+- Added the `runner` field in the config file.
+
 ## [1.0.4] - 2022-11-16
 ### Fixes
 - Fix to use correctly the function `[].concat([])` in dmark module, specially for apply the rest.